requirements.txt
easyocr
pdf2image
numpy
pillow
fastapi
uvicorn[standard]
google-auth
google-auth-httplib2
google-auth-oauthlib
requests
google-api-python-client
google-cloud-aiplatform
